Text editor bugfixes
Did not fix rare save glitch with temporary workaround. It is believed
to have been caused by the "clear_segment" function, which has been
removed. A bug when attempting to delete the end of the file causing the
cursor to act strangely has been fixed, and the code for the "up"
function has been cleaned up a little bit.